+++ app/src/main/java/com/bluehub/fastmixer/audio/MixingEngine.kt
@@ -32,7 +32,7 @@ class MixingEngine {
 
         @JvmStatic external fun getCurrentPlaybackProgress(): Int
 
-        @JvmStatic external fun gainSourceByDb(filePath: String, db: Int)
+        @JvmStatic external fun gainSourceByDb(filePath: String, db: Float)
 
         @JvmStatic external fun applySourceTransformation(filePath: String)
